Silenced Tableschema error when files have no extension
I spent quite a bit of time looking for a solution where we could
still process field detection by passing the tables directly into
tableschema, but didn't find a solution. This affects both files
without extensions, and generally files which cannot be read by
tableschema such as feather and parquet files. There should be a
way to do this, but I couldn't figure it out from the documentation.