Fixed upload exit codes to return unique values for various errors
Previously, upload simply returned an error if there was an error,
but it wasn't possible to tell which error it was. For anyone looking
to make bash scripts, this could be a problem checking for various
errors as it wouldn't be possible to tell them apart.

Additionally, added more debugging for various states in the
upload command.