Removed old required fields (or fixed bug if mimetype was undetermined)
Previously Pilot was still checking for minimum fields which should
be filled out by users. This was fixed earlier by automatically
filling in that information or not raising a validation error if it
were missing. The problem came when mimetype, a required field, could
not be determined and was left blank, which resulted in a raised error.

With the removal of 'required fields', pilot will now never throw errors
due to missing fields unless they come from jsonschema. (Mimetype will
also never be blank due to improved mimetype checking labeling a file
as either text or binary if a specific mimetype cannot be determined).