Silenced normal debug errors when listing
Many entries will lack a complete list of data to show when running
the `pilot list` command. This results in a TON of key error exceptions
in the backend which are caught and dealt with, but still logged. When
debugging is turned on, this becomes a mess.

This simply stops logging the exceptions.