Spec-change: Pilot SDK no longer raises exceptions for non-errors
Previously, DryRun and NoUpdatedNeeded exceptions would be raised
for usage with the SDK. This previosly matched the behavior with
the CLI, but would have caused headaches down the line for people
writing scripts and not catching those exact exceptions.

In addition, the PilotClient.upload() and .register() methods now
return a full listing of info instead of the newest metadata.
This lets users fetch info on everything from the new metadata,
the search ingest response info, the upload info respones, and
general metadata about the differences between the old and new
search records. This info is much better communicated with a
returned dict instead of raising an exception.