WIP fix for record 'collisions' for registering with multi-file recs
This is a work in progress for fixing a bug related to multi-file
collections. Previously, it was possible to upload new records
within multi-file collections. The new changes raise an error if
this is attempted, however, testing still seems to reveal a problem
where updating an existing record is falsly detected as a multi-file
collection, raising the new error when it should simply update the
record.

The API change includes a 'precise' flag to get_search_entry()
When True, get_search_entry() will only return the entry when
the `path` exactly matches a record or a file within a multi-file
record. When precise=False, any path within a multi-file record
will result in get_search_entry() returning the multi-file record.

Tests are still broken for these changes.