Set version on post_search queries
This will make pliot use the older style, prior to the newer style
coming out in a few days.

I started on switching Pilot over to the newer style, but it ended
up being a little more complex than I expected. Since we added
multi-file-entries, Pilot doesn't always get_subject when you ask
it to fetch a subject. Sometimes it needs to do a combination to
resolve whether you're talking about an exact search record, or a
file within a search record. The mechanisms for resolving entries
and files within entries are pretty well coupled to the older style.