fix: trailing slash causing dataset directory names to be ignored
This fixes a critical flaw where a trailing slash could cause a
directory to be ignored. For example, when uploading a folder
called 'foo/', the 'foo/' directory would be left out and files
would be uploaded to 'base_dir/', such as 'base_dir/file1.txt',
'base_dir/file2.txt'. Uploading files this way wasn't a problem,
however this made it possible to delete the base directory on the
next invocation of deleting the previously uploaded multi-file
dataset.